Add base code for multiple server (schools) support

+++ app/src/main/java/net/underdesk/circolapp/data/CircularRepository.kt
@@ -2,50 +2,49 @@ package net.underdesk.circolapp.data
 
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
-import net.underdesk.circolapp.server.DataFetcher
-import java.io.IOException
+import net.underdesk.circolapp.server.ServerAPI
 
 class CircularRepository(
     val circularDao: CircularDao,
-    private val fetcher: DataFetcher = DataFetcher()
+    private val serverAPI: ServerAPI
 ) {
     suspend fun updateCirculars(returnNewCirculars: Boolean = true): Pair<List<Circular>, Boolean> {
         var onlyNewCirculars = listOf<Circular>()
 
-        try {
-            withContext(Dispatchers.IO) {
-                val oldCirculars = circularDao.getCirculars()
-                val newCirculars = fetcher.getCircularsFromServer()
-                if (newCirculars.size != oldCirculars.size) {
-                    if (newCirculars.size < oldCirculars.size) {
-                        circularDao.deleteAll()
-                    }
+        return withContext(Dispatchers.IO) {
+            val result = serverAPI.getCircularsFromServer()
+            if (result.second == ServerAPI.Companion.Result.ERROR)
+                return@withContext Pair(emptyList(), false)
 
-                    if (returnNewCirculars) {
-                        val oldCircularsSize =
-                            if (newCirculars.size < oldCirculars.size) 0 else oldCirculars.size
+            val oldCirculars = circularDao.getCirculars()
+            val newCirculars = result.first
 
-                        val circularCount = newCirculars.size - oldCircularsSize
-                        onlyNewCirculars = newCirculars.subList(0, circularCount)
-                    }
-
-                    circularDao.insertAll(newCirculars)
+            if (newCirculars.size != oldCirculars.size) {
+                if (newCirculars.size < oldCirculars.size) {
+                    circularDao.deleteAll()
                 }
-            }
-        } catch (exception: IOException) {
-            return Pair(onlyNewCirculars, false)
-        }
 
-        return Pair(onlyNewCirculars, true)
+                if (returnNewCirculars) {
+                    val oldCircularsSize =
+                        if (newCirculars.size < oldCirculars.size) 0 else oldCirculars.size
+
+                    val circularCount = newCirculars.size - oldCircularsSize
+                    onlyNewCirculars = newCirculars.subList(0, circularCount)
+                }
+
+                circularDao.insertAll(newCirculars)
+            }
+            Pair(onlyNewCirculars, true)
+        }
     }
 
     companion object {
         @Volatile
         private var instance: CircularRepository? = null
 
-        fun getInstance(circularDao: CircularDao) =
+        fun getInstance(circularDao: CircularDao, serverAPI: ServerAPI) =
             instance ?: synchronized(this) {
-                instance ?: CircularRepository(circularDao).also { instance = it }
+                instance ?: CircularRepository(circularDao, serverAPI).also { instance = it }
             }
     }
 }
